Improve FindServers result handling
If FindServers returns, the list of discovery URLs is put into
a combo box and the first entry is selected by default. If the
first URL is not reachable, the user has to decide if the URL
used for the FindServers call should be attempted.
By looking for that URL in the returned discovery URLs list and
selecting it if available, this can be omitted in most cases.